requirements.txt
asgiref==3.6.0
Django==3.2.19
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
Pillow==9.5.0
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
whitenoise==6.4.0
